Extract the functions into the `funcs` package
