+++ opa/funcs/conversion_test.go
@@ -1,4 +1,4 @@
-package opa
+package funcs
 
 import (
 	"encoding/json"
@@ -9,6 +9,7 @@ import (
 	"github.com/hashicorp/hcl/v2/hclsyntax"
 	"github.com/terraform-linters/tflint-plugin-sdk/hclext"
 	"github.com/terraform-linters/tflint-plugin-sdk/tflint"
+	"github.com/terraform-linters/tflint-ruleset-opa/opa/tester"
 	"github.com/zclconf/go-cty/cty"
 )
 
@@ -97,7 +98,7 @@ func TestJSONToSchema(t *testing.T) {
 				}
 				return
 			}
-			if err == nil && test.err != "" {
+			if test.err != "" {
 				t.Fatal("should return an error, but it does not")
 			}
 
@@ -159,7 +160,7 @@ func TestJSONToOption(t *testing.T) {
 				}
 				return
 			}
-			if err == nil && test.err != "" {
+			if test.err != "" {
 				t.Fatal("should return an error, but it does not")
 			}
 
@@ -196,7 +197,7 @@ func TestTypedBlocksToJSON(t *testing.T) {
 		},
 	}
 
-	runner, diags := NewTestRunner(map[string]string{})
+	runner, diags := tester.NewRunner(map[string]string{})
 	if diags.HasErrors() {
 		t.Fatal(diags)
 	}
@@ -240,7 +241,7 @@ func TestNamedBlocksToJSON(t *testing.T) {
 		},
 	}
 
-	runner, diags := NewTestRunner(map[string]string{})
+	runner, diags := tester.NewRunner(map[string]string{})
 	if diags.HasErrors() {
 		t.Fatal(diags)
 	}
@@ -282,7 +283,7 @@ func TestBlocksToJSON(t *testing.T) {
 		},
 	}
 
-	runner, diags := NewTestRunner(map[string]string{})
+	runner, diags := tester.NewRunner(map[string]string{})
 	if diags.HasErrors() {
 		t.Fatal(diags)
 	}
@@ -328,7 +329,7 @@ func TestLocalsToJSON(t *testing.T) {
 		},
 	}
 
-	runner, diags := NewTestRunner(map[string]string{})
+	runner, diags := tester.NewRunner(map[string]string{})
 	if diags.HasErrors() {
 		t.Fatal(diags)
 	}
@@ -396,7 +397,7 @@ func TestBodyToJSON(t *testing.T) {
 		},
 	}
 
-	runner, diags := NewTestRunner(map[string]string{})
+	runner, diags := tester.NewRunner(map[string]string{})
 	if diags.HasErrors() {
 		t.Fatal(diags)
 	}
@@ -703,7 +704,7 @@ resource "aws_instance" "main" {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			runner, diags := NewTestRunner(map[string]string{"main.tf": test.source})
+			runner, diags := tester.NewRunner(map[string]string{"main.tf": test.source})
 			if diags.HasErrors() {
 				t.Fatal(diags)
 			}
@@ -752,7 +753,7 @@ func TestNestedBlockToJSON(t *testing.T) {
 		},
 	}
 
-	runner, diags := NewTestRunner(map[string]string{})
+	runner, diags := tester.NewRunner(map[string]string{})
 	if diags.HasErrors() {
 		t.Fatal(diags)
 	}
@@ -762,72 +763,6 @@ func TestNestedBlockToJSON(t *testing.T) {
 			got, err := nestedBlockToJSON(test.input, map[string]cty.Type{}, "", runner)
 			if err != nil {
 				t.Fatal(err)
-			}
-
-			if diff := cmp.Diff(test.want, got); diff != "" {
-				t.Error(diff)
-			}
-		})
-	}
-}
-
-func TestJSONToIssue(t *testing.T) {
-	tests := []struct {
-		name  string
-		input any
-		want  *Issue
-		err   string
-	}{
-		{
-			name: "valid issue",
-			input: map[string]any{
-				"msg": "message",
-				"range": map[string]any{
-					"filename": "",
-					"start":    map[string]any{"line": json.Number("0"), "column": json.Number("0"), "byte": json.Number("0")},
-					"end":      map[string]any{"line": json.Number("0"), "column": json.Number("0"), "byte": json.Number("0")},
-				},
-			},
-			want: &Issue{Message: "message"},
-		},
-		{
-			name:  "invalid type",
-			input: "",
-			err:   "issue is not object, got string",
-		},
-		{
-			name: "invalid message type",
-			input: map[string]any{
-				"msg": 1,
-				"range": map[string]any{
-					"filename": "",
-					"start":    map[string]any{"line": json.Number("0"), "column": json.Number("0"), "byte": json.Number("0")},
-					"end":      map[string]any{"line": json.Number("0"), "column": json.Number("0"), "byte": json.Number("0")},
-				},
-			},
-			err: "issue.msg is not string, got int",
-		},
-		{
-			name: "invalid range type",
-			input: map[string]any{
-				"msg":   "test",
-				"range": "main.tf:1,1-1",
-			},
-			err: "issue.range is not object, got string",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			got, err := jsonToIssue(test.input, "issue")
-			if err != nil {
-				if err.Error() != test.err {
-					t.Fatalf(`expect "%s", but got "%s"`, test.err, err.Error())
-				}
-				return
-			}
-			if err == nil && test.err != "" {
-				t.Fatal("should return an error, but it does not")
 			}
 
 			if diff := cmp.Diff(test.want, got); diff != "" {
@@ -965,7 +900,7 @@ func TestJSONToRange(t *testing.T) {
 				}
 				return
 			}
-			if err == nil && test.err != "" {
+			if test.err != "" {
 				t.Fatal("should return an error, but it does not")
 			}
